+++ packages/gov-ui-kit/src/core/components/tabs/tabsContent/tabsContent.tsx
@@ -1,10 +1,15 @@
-import { TabsContent as RadixTabsContent, type TabsContentProps as RadixTabsContentProps } from '@radix-ui/react-tabs';
+import { TabsContent as RadixTabsContent } from '@radix-ui/react-tabs';
+import { type ComponentProps } from 'react';
 
-export interface ITabsContentProps extends RadixTabsContentProps {
+export interface ITabsContentProps extends ComponentProps<'div'> {
     /**
      * Value linking Tabs.Content to its corresponding Tabs.Trigger
      */
     value: string;
+    /**
+     * When `true`, the content will stay mounted even when inactive.
+     */
+    forceMount?: true;
 }
 
 export const TabsContent: React.FC<ITabsContentProps> = (props) => {