+++ packages/gov-ui-kit/src/core/components/tabs/tabsRoot/tabsRoot.tsx
@@ -1,11 +1,19 @@
-import { Tabs as RadixTabsRoot, type TabsProps as RadixTabsProps } from '@radix-ui/react-tabs';
-import { createContext, forwardRef, useMemo } from 'react';
+import { Tabs as RadixTabsRoot } from '@radix-ui/react-tabs';
+import { createContext, forwardRef, useMemo, type ComponentPropsWithoutRef } from 'react';
 
-export interface ITabsRootProps extends Omit<RadixTabsProps, 'orientation'> {
+export interface ITabsRootProps extends ComponentPropsWithoutRef<typeof RadixTabsRoot> {
     /**
      * The value of the tab that should be selected by default.
      */
     defaultValue?: string;
+    /**
+     * The value of the selected tab.
+     */
+    value?: string;
+    /**
+     * Callback when the value changes.
+     */
+    onValueChange?: (value: string) => void;
     /**
      * Whether the Tabs.List should use an underlined style. @default false
      */
@@ -22,14 +30,21 @@ export interface ITabsContext {
 export const TabsContext = createContext<ITabsContext>({ isUnderlined: false });
 
 export const TabsRoot = forwardRef<HTMLDivElement, ITabsRootProps>((props, ref) => {
-    const { children, isUnderlined = false, ...otherProps } = props;
+    const { children, isUnderlined = false, defaultValue, value, onValueChange, ...otherProps } = props;
     const tabsContextValue = useMemo(() => ({ isUnderlined }), [isUnderlined]);
 
     return (
-        <RadixTabsRoot ref={ref} orientation="horizontal" {...otherProps}>
+        <RadixTabsRoot
+            ref={ref}
+            defaultValue={defaultValue}
+            value={value}
+            onValueChange={onValueChange}
+            orientation="horizontal"
+            {...otherProps}
+        >
             <TabsContext.Provider value={tabsContextValue}>{children}</TabsContext.Provider>
         </RadixTabsRoot>
     );
 });
 
-TabsRoot.displayName = 'Tabs.Root';
+TabsRoot.displayName = 'TabsRoot';